tasks: rename tags and update the test
