+++ test/tasks/js/expected/js-in-template.js
