+++ test/tasks/js/src/js/js-in-template.js
