+++ test/tasks/js/src/template/example1.njk
@@ -1,3 +1,3 @@
->>>js
+<script to='bp-js'>
 console.log('JS in template 1');
->>>endjs
+</script>
