+++ test/tasks/js/src/template/example2.njk
@@ -1,3 +1,3 @@
->>>js
+<script to='bp-js'>
 console.log('JS in template 2');
->>>endjs
+</script>
