+++ test/tasks/sass/expected/_sass-in-template.scss
