+++ test/tasks/sass/src/sass/index.scss
@@ -2,4 +2,4 @@
 @import "example";
 
 // Required
-@import "auto-extract";
+@import "sass-in-template";
