+++ test/tasks/sass/src/template/example1.njk
@@ -1,6 +1,6 @@
->>>sass
+<style to='bp-sass'>
 .sass-in-template1 {
   display: block;
   color: $color;
 }
->>>endsass
+</style>
