+++ test/tasks/sass/src/template/example2.njk
@@ -1,6 +1,6 @@
->>>sass
+<style to='bp-sass'>
 .sass-in-template2 {
   display: block;
   color: $color;
 }
->>>endsass
+</style>
