+++ test/tasks/skin/expected/skin-in-template.css
