+++ test/tasks/skin/src/skin/index.css
@@ -2,4 +2,4 @@
 @import "example";
 
 /* Required */
-@import "auto-extract";
+@import "skin-in-template";
