+++ test/tasks/skin/src/skin/skin-in-template.css
