+++ test/tasks/skin/src/template/example1.njk
@@ -1,6 +1,6 @@
->>>skin
+<style to='bp-skin'>
 .skin-in-template1 {
   display: block;
   color: $(color);
 }
->>>endskin
+</style>
