+++ test/tasks/skin/src/template/example2.njk
@@ -1,6 +1,6 @@
->>>skin
+<style to='bp-skin'>
 .skin-in-template2 {
   display: block;
   color: $(color);
 }
->>>endskin
+</style>
