+++ test/tasks/template/src/index.njk
@@ -1,11 +1,11 @@
->>>template
-{% extends "layout.njk" %}
+<template to='bp-template'>
+  {% extends "layout.njk" %}
 
-{% block main %}
-{% template "template/template-main.njk" %}
-{% endblock %}
+  {% block main %}
+  {% template "template/template-main.njk" %}
+  {% endblock %}
 
-{% block sidebar %}
-{% template "template/template-sidebar.njk" %}
-{% endblock %}
->>>endtemplate
+  {% block sidebar %}
+  {% template "template/template-sidebar.njk" %}
+  {% endblock %}
+</template>
