+++ test/tasks/template/src/layout.njk
@@ -1,11 +1,11 @@
->>>template
-<div>
-  <div class='main'>
-    {% block main %}{% endblock %}
-  </div>
+<template to='bp-template'>
+  <div>
+    <div class='main'>
+      {% block main %}{% endblock %}
+    </div>
 
-  <div class='sidebar'>
-    {% block sidebar %}{% endblock %}
+    <div class='sidebar'>
+      {% block sidebar %}{% endblock %}
+    </div>
   </div>
-</div>
->>>endtemplate
+</template>
