+++ test/tasks/template/src/template/template-main.njk
@@ -1,7 +1,7 @@
->>>template
-<div>
-  <div>Main content</div>
-  <div>x = {{ data.x }}</div>
-  <div>Package name = {{ pkg.name }}</div>
-</div>
->>>endtemplate
+<template to='bp-template'>
+  <div>
+    <div>Main content</div>
+    <div>x = {{ data.x }}</div>
+    <div>Package name = {{ pkg.name }}</div>
+  </div>
+</template>
