+++ test/tasks/template/src/template/template-sidebar.njk
@@ -1,5 +1,5 @@
->>>template
-<div>
-  <div>Sidebar content</div>
-</div>
->>>endtemplate
+<template to='bp-template'>
+  <div>
+    <div>Sidebar content</div>
+  </div>
+</template>
